// src/app.common/localization/generated/user.enum.ts
// This file is auto-generated. Do not edit manually.
export enum UserKeys {
  FRIEND_REQUEST_SENT = 'FRIEND_REQUEST_SENT',
  FRIEND_REMOVED = 'FRIEND_REMOVED',
  TEAM_INVITATION_SENT = 'TEAM_INVITATION_SENT',
  TEAM_INVITATION_CANCELLED = 'TEAM_INVITATION_CANCELLED',
  REMOVED_FROM_TEAM = 'REMOVED_FROM_TEAM',
  USER_NOW_CHIEF = 'USER_NOW_CHIEF',
  FRIEND_REQUEST_ACCEPTED = 'FRIEND_REQUEST_ACCEPTED',
  TEAM_INVITATION_ACCEPTED = 'TEAM_INVITATION_ACCEPTED',
  INVALID_ACTION = 'INVALID_ACTION',
  USER_INFO_UPDATED = 'USER_INFO_UPDATED',
  OTP_RESENT = 'OTP_RESENT',
  ADD_TO_FRIENDS = 'ADD_TO_FRIENDS',
  REMOVE_FROM_FRIENDS = 'REMOVE_FROM_FRIENDS',
  ADD_TO_TEAM = 'ADD_TO_TEAM',
  REMOVE_FROM_TEAM = 'REMOVE_FROM_TEAM',
  MAKE_CHIEF = 'MAKE_CHIEF',
  ACCEPT_FRIEND_REQUEST = 'ACCEPT_FRIEND_REQUEST',
  ACCEPT_TEAM_REQUEST = 'ACCEPT_TEAM_REQUEST',
  OTP_SENT_FOR_EMAIL_CHANGE = 'OTP_SENT_FOR_EMAIL_CHANGE',
  SENDING_OTP = 'SENDING_OTP',
  TEAM_INVITATION_ALREADY_SENT = 'TEAM_INVITATION_ALREADY_SENT',
  NO_FRIEND_REQUEST = 'NO_FRIEND_REQUEST',
  NO_TEAM_INVITATION = 'NO_TEAM_INVITATION',
  INCORRECT_OTP = 'INCORRECT_OTP',
  EMAIL_UPDATED = 'EMAIL_UPDATED',
  REQUEST_NOT_FOUND = 'REQUEST_NOT_FOUND',
  STRANGER = 'STRANGER',
  FRIENDS = 'FRIENDS',
  TEAMMATES = 'TEAMMATES',
  FRIENDS_TEAMMATES = 'FRIENDS_TEAMMATES'
}